+++ core-ng/src/main/java/core/framework/internal/http/RetryInterceptor.java
@@ -3,7 +3,6 @@ package core.framework.internal.http;
 import core.framework.api.http.HTTPStatus;
 import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogManager;
-import core.framework.log.ActionLogContext;
 import okhttp3.Interceptor;
 import okhttp3.Request;
 import okhttp3.Response;
@@ -38,6 +37,7 @@ public class RetryInterceptor implements Interceptor {
         Request request = chain.request();
         int attempts = 1;
         while (true) {
+            if (chain.call().isCanceled()) throw new IOException("call timeout");   // AsyncTimout cancels call if callTimeout, refer to RealCall.kt/timout field
             try {
                 Response response = chain.proceed(request);
                 int statusCode = response.code();
@@ -56,7 +56,10 @@ public class RetryInterceptor implements Interceptor {
             }
             sleep.sleep(waitTime(attempts));
             attempts++;
-            ActionLogContext.stat("http_retries", 1);
+
+            // set to actionLog directly to keep trace log concise
+            ActionLog actionLog = LogManager.CURRENT_ACTION_LOG.get();
+            if (actionLog != null) actionLog.stats.compute("http_retries", (key, oldValue) -> (oldValue == null) ? 1.0 : oldValue + 1);
         }
     }
 
@@ -83,7 +86,7 @@ public class RetryInterceptor implements Interceptor {
 
     boolean shouldRetry(int attempts, int statusCode) {
         return attempts < maxRetries
-                && (statusCode == HTTPStatus.SERVICE_UNAVAILABLE.code || statusCode == HTTPStatus.TOO_MANY_REQUESTS.code);
+               && (statusCode == HTTPStatus.SERVICE_UNAVAILABLE.code || statusCode == HTTPStatus.TOO_MANY_REQUESTS.code);
     }
 
     boolean shouldRetry(int attempts, String method, IOException e) {